requirements.txt
Pillow==8.1.0
tabulate==0.8.7
